onoff-n-release-control-dfss
Controlling the power through the grey wire connected to a relay and of
the quick release valve have been added to the code and tested on the
prototype. Tilt motor has been tested and values of motion range
adjusted. Format of message to GS has been modified to accomodate for a
value of "0" when the distance sensors are not connected. Comments on
what to watch out for for test in Gamesa.